+++ next/src/ui/footer/index.tsx
@@ -1,11 +1,11 @@
 import { getSite } from '@/lib/sanity'
-import Social from '../Social'
+import Social from '@/ui/Social'
 
 export default async function Footer() {
 	const { title, footerMenu } = await getSite()
 
 	return (
-		<footer className="bg-black p-4 text-center text-white">
+		<footer className="bg-ink text-canvas p-4 text-center">
 			<div className="mx-auto max-w-screen-xl space-y-4">
 				<Social className="justify-center" />