+++ next/src/ui/header/LinkList.tsx
@@ -6,7 +6,7 @@ export default function LinkList({ label, links }: Sanity.LinkList) {
 		<InteractiveDetails className="relative" closeAfterNavigate>
 			<summary>{label}</summary>
 
-			<ul className="anim-fade-to-b left-0 top-full border p-2 md:absolute md:min-w-max md:bg-white/90 md:backdrop-blur">
+			<ul className="anim-fade-to-b md:bg-canvas/90 left-0 top-full border p-2 md:absolute md:min-w-max md:backdrop-blur">
 				{links?.map((link, key) => (
 					<li key={key}>
 						<CTA className="link" link={link} />