+++ next/src/ui/header/index.tsx
@@ -2,7 +2,7 @@ import { getSite } from '@/lib/sanity'
 import Wrapper from './Wrapper'
 import Link from 'next/link'
 import Menu from './Menu'
-import CTAList from '../CTAList'
+import CTAList from '@/ui/CTAList'
 import Toggle from './Toggle'
 import { cn } from '@/lib/utils'
 import css from './Header.module.css'
@@ -11,7 +11,7 @@ export default async function Header() {
 	const { title, ctas } = await getSite()
 
 	return (
-		<Wrapper className="sticky top-0 z-10 border-b bg-white/90 backdrop-blur">
+		<Wrapper className="bg-canvas/90 border-ink/10 sticky top-0 z-10 border-b backdrop-blur">
 			<div
 				className={cn(
 					css.header,